pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-belief-narrative-pipeline"
version = "0.1.0"
description = "Local-first, S3-compatible data pipeline for crypto belief + narrative + underreaction research."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Tristan Agaesse" }]
dependencies = [
  "boto3>=1.34",
  "botocore>=1.34",
  "dagster>=1.7",
  "dagster-postgres>=0.23",
  "dagster-duckdb>=0.23",
  "dagster-docker>=0.23",
  "dagster-webserver>=1.7",
  "duckdb>=1.0.0",
  "gdeltdoc>=1.0",
  "httpx>=0.27",
  "pandas>=2.2",
  "polars>=1.0.0",
  "pyarrow>=16.0.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "PyYAML>=6.0",
  "soda-core-duckdb>=3.4",
  "tenacity>=8.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
crypto_belief_pipeline = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
